refactor: cleaned up the code a bit
