+++ src/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import { withAuthentication } from "./components/Auth";
 
 import "./App.scss";
-import SearchHeader from "./components/searchheader.js";
+import SearchView from "./components/searchview.js";
 import LoginPage from "./components/loginpage.js";
 import PrivacyView from "./components/privacyview";
 //import GoogleApiWrapper from './components/routemap.js';
@@ -14,7 +14,7 @@ import * as ROUTES from "./constants/routes";
 const App = () => (
   <Router>
     <Switch>
-      <Route exact path={ROUTES.LANDING} component={SearchHeader} />
+      <Route exact path={ROUTES.LANDING} component={SearchView} />
       <Route exact path={ROUTES.LOGIN} component={LoginPage} />
       <Route exact path={ROUTES.PRIVACY} component={PrivacyView} />
       {/* <Route component={NoMatch} /> */}
